feat: implement user identification in comments and posts, enhance like status handling

+++ src/blogger-platform/comments/application/handlers/delete-comment.command-handler.ts
@@ -1,5 +1,5 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
-import { NotFoundException } from '@nestjs/common';
+import { ForbiddenException, NotFoundException } from '@nestjs/common';
 import { DeleteCommentCommand } from '../commands/delete-comment.command';
 import { CommentsRepository } from '../../comments.repository';
 
@@ -8,12 +8,16 @@ export class DeleteCommentCommandHandler implements ICommandHandler<DeleteCommen
   constructor(private readonly commentsRepository: CommentsRepository) {}
 
   async execute(command: DeleteCommentCommand): Promise<void> {
-    const { commentId } = command;
+    const { commentId, userId } = command;
 
     const comment = await this.commentsRepository.findOne(commentId);
 
     if (!comment) {
       throw new NotFoundException(`No such comment with id: ${commentId}`);
+    }
+
+    if (comment.commentatorInfo.userId !== userId) {
+      throw new ForbiddenException();
     }
 
     await this.commentsRepository.deleteComment(commentId);
